refactor: move commons utilities inside common and bootstrap oneid-lambda-service-metadata
