+++ src/oneid/oneid-ecs-core/src/main/java/it/pagopa/oneid/common/SAMLUtilsConstants.java
@@ -1,4 +1,4 @@
-package it.pagopa.oneid.service.utils;
+package it.pagopa.oneid.common;
 
 import org.eclipse.microprofile.config.inject.ConfigProperty;
