+++ src/oneid/oneid-ecs-core/src/main/java/it/pagopa/oneid/service/SAMLService.java
@@ -1,9 +1,9 @@
 package it.pagopa.oneid.service;
 
+import it.pagopa.oneid.common.exception.OneIdentityException;
+import it.pagopa.oneid.common.exception.SAMLUtilsException;
 import it.pagopa.oneid.exception.GenericAuthnRequestCreationException;
 import it.pagopa.oneid.exception.IDPSSOEndpointNotFoundException;
-import it.pagopa.oneid.exception.OneIdentityException;
-import it.pagopa.oneid.exception.SAMLUtilsException;
 import it.pagopa.oneid.exception.SAMLValidationException;
 import java.util.List;
 import java.util.Optional;
@@ -28,6 +28,6 @@ public interface SAMLService {
   Optional<EntityDescriptor> getEntityDescriptorFromEntityID(String entityID)
       throws SAMLUtilsException;
 
-  void checkSAMLStatus(Response response) throws OneIdentityException;
+  void checkSAMLStatus(Response response) throws OneIdentityException, OneIdentityException;
 
 }
