+++ src/oneid/oneid-lambda-service-metadata/src/main/java/it/pagopa/oneid/ServiceMetadata.java
@@ -4,12 +4,13 @@ import jakarta.ws.rs.GET;
 import jakarta.ws.rs.Path;
 import jakarta.ws.rs.core.Response;
 
-@Path("/")
+@Path("/saml")
 public class ServiceMetadata {
 
   @GET
   @Path("/metadata")
   public Response metadata() {
+    
     return Response.ok("Metadata path").build();
   }
 }
